Add wsgi.py and take out /api from routes.
Since we're using an app factory, we need a python file to call the factory and also to reference via uwsgi.

Also remove /api from all the endpoints.  We will mount all the backend endpoints to /api directly on the nginx server config file.  This was the easier path vs mounting to the same endpoint as the react application.